+++ raster/r.in.mat/main.c
@@ -212,11 +212,12 @@ int main(int argc, char *argv[])
                 G_fatal_error(_("Invalid 'map_name' array"));
 
             if (data_format == 5) {
-                if (fread(&map_name, sizeof(char), ncols, fp1) != ncols)
+                if (fread(&map_name, sizeof(char), ncols, fp1) != (size_t)ncols)
                     G_fatal_error(_("Error reading data"));
             }
             else if (data_format == 0) { /* sigh.. */
-                if (fread(&map_name_d, sizeof(double), ncols, fp1) != ncols)
+                if (fread(&map_name_d, sizeof(double), ncols, fp1) !=
+                    (size_t)ncols)
                     G_fatal_error(_("Error reading data"));
                 for (i = 0; i < ncols; i++)
                     map_name[i] = (char)map_name_d[i];
@@ -271,12 +272,13 @@ int main(int argc, char *argv[])
                 G_fatal_error(_("Invalid 'map_title' array"));
 
             if (data_format == 5) {
-                if (fread(&map_title, sizeof(char), ncols, fp1) != ncols)
+                if (fread(&map_title, sizeof(char), ncols, fp1) !=
+                    (size_t)ncols)
                     G_fatal_error(_("Error reading data"));
             }
             else if (data_format == 0) { /* sigh.. */
                 if (fread(&map_name_d, sizeof(double), ncols, fp1) !=
-                    ncols) /* note reusing variable */
+                    (size_t)ncols) /* note reusing variable */
                     G_fatal_error(_("Error reading data"));
                 for (i = 0; i < ncols; i++)
                     map_title[i] = (char)map_name_d[i];
@@ -306,7 +308,7 @@ int main(int argc, char *argv[])
                 array_data =
                     G_calloc(mrows * (ncols + 1), Rast_cell_size(map_type));
                 if (fread(array_data, sizeof(double), (size_t)mrows * ncols,
-                          fp1) != (mrows * ncols))
+                          fp1) != (size_t)(mrows * ncols))
                     G_fatal_error(_("Error reading data"));
                 break;
             case 1:
@@ -315,7 +317,7 @@ int main(int argc, char *argv[])
                 array_data =
                     G_calloc(mrows * (ncols + 1), Rast_cell_size(map_type));
                 if (fread(array_data, sizeof(float), (size_t)mrows * ncols,
-                          fp1) != (mrows * ncols))
+                          fp1) != (size_t)(mrows * ncols))
                     G_fatal_error(_("Error reading data"));
                 break;
             case 2:
@@ -324,7 +326,7 @@ int main(int argc, char *argv[])
                 array_data =
                     G_calloc(mrows * (ncols + 1), Rast_cell_size(map_type));
                 if (fread(array_data, sizeof(int), (size_t)mrows * ncols,
-                          fp1) != (mrows * ncols))
+                          fp1) != (size_t)(mrows * ncols))
                     G_fatal_error(_("Error reading data"));
                 break;
             default:
